Add Travis CI support
Use Travis CI to deploy bundled versions of the micro site to both
staging and production with the AWS CLI (encapsulated in `cipublish`).
This uses a single set of credentials that are whitelisted to write to
both the staging and production buckets.

Also, adjust the interaction between CloudFront and S3 so that only S3
is able to read objects in the origin via an origin access policy.