Color popover arrows accordingly
When a popover is shown below the thumbnail, the arrow
is aligned with the title which has a prominent, non-
white background color. A plain white arrow looks
starkly out of place.

To fix this, we add the NLCD / Soil group name to the
arrow, and if and only if the popover is shown below
the thumbnail, we use the title color's arrow. Else,
we show the regular white arrow, which goes better
with the popover body.