Realign popovers to not obscure main content
Previously all the popovers would open to the left of
the thumbnail, uniformly. This unfortunately occluded
the main content. To avoid that, now we open the popover
of the left-most thumbnail of every row to the right.
For all the NLCD thumbnails, the popover is opened
below them. For the Soil thumbnails, the popover is
opened above.

This works well for all cases except the right-most
ones, whose content is a little squished for lack of
room. This may be an acceptable compromise given the
client's needs.

Refs #19